requirements.txt
selenium==4.2.0
pyunitreport==0.1.4
python-dotenv==0.20.0
black==22.3.0
flake8==4.0.1
